Begin with a proper CLI
