+++ Cargo.toml
@@ -82,6 +82,7 @@ noise = { version = "0.7", default-features = false }
 obj = "0.10"
 # parking_lot 0.12 switches from `winapi` to `windows`; permit either
 parking_lot = ">=0.11,<0.13"
+pico-args = { version = "0.5.0", features = ["eq-separator", "short-space-opt", "combined-flags"] }
 png = "0.17.8"
 pollster = "0.2"
 profiling = { version = "1", default-features = false }
